requirements.txt
flask
flask_cors
google
google.genai
pillow